Frontend: more updates to splash (still testing)
