+++ frontend/android/app/src/main/res/values-night/colors.xml
@@ -1,3 +1,1 @@
-<resources>
-  <color name="splashscreen_background">#c41111</color>
-</resources>
+<resources/>
