+++ frontend/android/app/src/main/res/values/colors.xml
@@ -3,4 +3,5 @@
   <color name="iconBackground">#ffffff</color>
   <color name="colorPrimary">#023c69</color>
   <color name="colorPrimaryDark">#c41111</color>
+  <color name="activityBackground">#c41111</color>
 </resources>
