+++ frontend/android/app/src/main/res/values/styles.xml
@@ -2,6 +2,7 @@
   <style name="AppTheme" parent="Theme.EdgeToEdge">
     <item name="android:editTextBackground">@drawable/rn_edit_text_material</item>
     <item name="colorPrimary">@color/colorPrimary</item>
+    <item name="android:windowBackground">@color/activityBackground</item>
     <item name="android:statusBarColor">#c41111</item>
   </style>
   <style name="Theme.App.SplashScreen" parent="Theme.SplashScreen">
