+++ frontend/app/_layout.tsx
@@ -6,13 +6,13 @@ export default function RootLayout() {
   return (
     <View className="flex-1 bg-white">
       <Stack
-        initialRouteName="splash"
         screenOptions={{
           headerShown: false,
         }}
       >
-        <Stack.Screen name="splash" options={{ animation: "fade" }} />
-        <Stack.Screen name="index" />
+        {/* No splash screen route needed; index.tsx is now the splash */}
+        <Stack.Screen name="index" options={{ animation: "fade" }} />
+        <Stack.Screen name="home" />
         <Stack.Screen name="signup" />
       </Stack>
     </View>
